modificando cartAPI para clean architecture

+++ src/Core/TecnoMundo.Application/DTOs/CouponVO.cs
@@ -4,6 +4,6 @@
     {
         public Guid Id { get; set; }
         public string CouponCode { get; set; }
-        public float DiscountAmount { get; set; }
+        public decimal DiscountAmount { get; set; }
     }
 }
